+++ crates/knok-core/src/ast.rs
@@ -68,6 +68,10 @@ pub enum CallOp {
     Exp,
     Greater,
     GreaterEqual,
+    Gather {
+        target: TensorType,
+        axis: usize,
+    },
     IsNan,
     Less,
     LessEqual,
@@ -103,6 +107,9 @@ pub enum CallOp {
     Take {
         axis: usize,
         index: usize,
+    },
+    TakeAlongAxis {
+        axis: usize,
     },
     Transpose,
     Unsqueeze(TensorType),